Created BlogTechnologyViewsUpdater processor
